Added logic validation on CRUD Users. (Business logic).

+++ Backend/UsersControl/Controllers/UsersController.cs
@@ -58,11 +58,17 @@ namespace UsersControl.Controllers
         public async Task<ActionResult<UserReadDTO>> CreateUser(UserCreateDTO createUserDTO)
         {
             var userModel = mapper.Map<User>(createUserDTO);
-            await modelService.CreateUser(userModel);
+            var errorMessage = modelService.ValidateBeforeCreate(userModel);
+            if (string.IsNullOrEmpty(errorMessage))
+            {
+                await modelService.CreateUser(userModel);
+                var userReadDTO = mapper.Map<UserReadDTO>(userModel);
 
-            var userReadDTO = mapper.Map<UserReadDTO>(userModel);
+                return CreatedAtRoute(nameof(GetUserById), new { Id = userReadDTO.Id }, userReadDTO);
+            }
 
-            return CreatedAtRoute(nameof(GetUserById), new { Id = userReadDTO.Id }, userReadDTO);
+            ModelState.AddModelError("Errors", errorMessage);
+            return BadRequest(ModelState);
         }
 
 
@@ -70,11 +76,19 @@ namespace UsersControl.Controllers
         public async Task<ActionResult> UpdateUser(int id, UserUpdateDTO updateUserDTO)
         {
             var userModel = await modelService.GetUserById(id);
+            var errorMessage = string.Empty;
             if (userModel != null)
             {
-                mapper.Map<UserUpdateDTO, User>(updateUserDTO, userModel);
-                await modelService.UpdateUser(userModel);
-                return NoContent();
+                errorMessage = modelService.ValidateBeforeUpdate(userModel);
+                if (string.IsNullOrEmpty(errorMessage))
+                {
+                    mapper.Map<UserUpdateDTO, User>(updateUserDTO, userModel);
+                    await modelService.UpdateUser(userModel);
+                    return NoContent();
+                }
+
+                ModelState.AddModelError("Errors", errorMessage);
+                return BadRequest(ModelState);
             }
 
             return NotFound();
@@ -85,10 +99,18 @@ namespace UsersControl.Controllers
         public async Task<ActionResult> DeleteUser(int id)
         {
             var userModel = await modelService.GetUserById(id);
+            var errorMessage = string.Empty;
             if (userModel != null)
             {
-                await modelService.DeleteUser(userModel);
-                return NoContent();
+                errorMessage = modelService.ValidateBeforeDelete(userModel);
+                if (string.IsNullOrEmpty(errorMessage))
+                {
+                    await modelService.DeleteUser(userModel);
+                    return NoContent();
+                }
+                
+                ModelState.AddModelError("Errors", errorMessage);
+                return BadRequest(ModelState);
             }
 
             return NotFound();
